+++ flightdataadapter.h
@@ -10,28 +10,28 @@
 class FlightDataAdapter : public QObject
 {
     Q_OBJECT
-    Q_PROPERTY(int TimeBootMs READ TimeBootMs NOTIFY flightDataChanged)
-    Q_PROPERTY(int Lat READ Lat NOTIFY flightDataChanged)
-    Q_PROPERTY(int Lon READ Lon NOTIFY flightDataChanged)
-    Q_PROPERTY(int Alt READ Alt NOTIFY flightDataChanged)
-    Q_PROPERTY(int RelativeAlt READ RelativeAlt NOTIFY flightDataChanged)
-    Q_PROPERTY(int Vx READ Vx NOTIFY flightDataChanged)
-    Q_PROPERTY(int Vy READ Vy NOTIFY flightDataChanged)
-    Q_PROPERTY(int Vz READ Vz NOTIFY flightDataChanged)
-    Q_PROPERTY(int Hdg READ Hdg NOTIFY flightDataChanged)
+    Q_PROPERTY(double TimeBootMs READ TimeBootMs NOTIFY flightDataChanged)
+    Q_PROPERTY(double Lat READ Lat NOTIFY flightDataChanged)
+    Q_PROPERTY(double Lon READ Lon NOTIFY flightDataChanged)
+    Q_PROPERTY(double Alt READ Alt NOTIFY flightDataChanged)
+    Q_PROPERTY(double RelativeAlt READ RelativeAlt NOTIFY flightDataChanged)
+    Q_PROPERTY(double Vx READ Vx NOTIFY flightDataChanged)
+    Q_PROPERTY(double Vy READ Vy NOTIFY flightDataChanged)
+    Q_PROPERTY(double Vz READ Vz NOTIFY flightDataChanged)
+    Q_PROPERTY(double Hdg READ Hdg NOTIFY flightDataChanged)
 public:
     explicit FlightDataAdapter(QObject *parent = nullptr);
     void SetFlightData(FlightData newData);
 
-    int TimeBootMs() const;
-    int Lat () const;
-    int Lon () const;
-    int Alt () const;
-    int RelativeAlt () const;
-    int Vx () const;
-    int Vy () const;
-    int Vz () const;
-    int Hdg () const;
+    double TimeBootMs() const;
+    double Lat () const;
+    double Lon () const;
+    double Alt () const;
+    double RelativeAlt () const;
+    double Vx () const;
+    double Vy () const;
+    double Vz () const;
+    double Hdg () const;
 
 signals:
     void flightDataChanged();
